Add input sanitization setup
This adds a simple sanitize_obj fn that can act on the k8s_obj that each
supported model can act on.

Unfortunately, this can only be done when we have ownership of the
object, so this should be done at input time.

Thus, we have restructured the whole input pipeline to use:

- From<k8s_openapi::SomeType> for OurKubeType where possible
- remove `KubeResouce::from_api` from the trait
- made all `from` impls take ownership of the objects

Note there was no reason not to take ownership. It only required moving
to `IntoIterator` over `Iterator` and with that change we were able to
use the new `sanitize_obj` fn during the `From::from` call.

Signed-off-by: clux <sszynrae@gmail.com>